+++ package.json
@@ -8,8 +8,12 @@
   },
   "dependencies": {
     "body-parser": "^1.19.0",
+    "cookie-session": "^1.4.0",
+    "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "mongodb": "^4.1.2"
+    "mongodb": "^4.1.2",
+    "node-fetch": "2.6.3",
+    "oauth": "^0.9.15"
   }
 }
