--- public/index.css
@@ -1,75 +1,0 @@
-/* Downloads Staatliches (https://fonts.google.com/specimen/Staatliches?sort=popularity) */
-@import url("https://fonts.googleapis.com/css2?family=Staatliches&display=swap");
-
-body {
-  /*
-    Colors from Wheel:
-    ##4C6670 - Dark blue/gray
-    #5CC8F2 - Robin egg blue
-    #A3DAF0 - Pale blue
-    #2B5C70 - Deep turquoise
-    #80ACBD - Light gray
-    */
-  --gradient-color: #80acbd;
-  --h1h2-color: #2b5c70;
-  --h2-border-color: #80acbd;
-  --h3-color: #5cc8f2;
-  --aLot-color: #5cc8f2;
-  --table-color: #2b5c70;
-}
-
-h1 {
-  text-align: center;
-  font-family: "Staatliches", cursive;
-  color: var(--h1h2-color);
-}
-
-h1 {
-  font-size: 300%;
-}
-
-label,
-table {
-  color: var(--h1h2-color);
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-}
-
-table {
-  background-color: white;
-  border: 0.2rem solid var(--table-color);
-}
-
-th {
-  background-color: lightgrey;
-}
-
-button {
-  margin: 0.2rem;
-}
-
-.homework-list-group,
-.newHomeworkPrompt {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  flex: 1 400px;
-}
-
-#overlay {
-  position: fixed;
-  display: none;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0,0,0,0.25); /* Black background with opacity */
-}
-
-.newHomeworkPrompt {
-  background-color: white;
-  margin: 20% auto;
-  width: 35%;
-  box-shadow: 0 6px 6px 0 rgba(0,0,0,0.15);
-}
